+++ src/black_holes/EAGLE/black_holes_part.h
@@ -31,7 +31,7 @@
 struct bpart {
 
   /*! Particle ID. */
-  long long id;
+  int64_t id;
 
   /*! Pointer to corresponding gravity part. */
   struct gpart* gpart;
@@ -154,13 +154,13 @@ struct bpart {
   int num_ngb_density;
 
   /*! List of interacting particles in the density SELF and PAIR */
-  long long ids_ngbs_density[MAX_NUM_OF_NEIGHBOURS_BLACK_HOLES];
+  int64_t ids_ngbs_density[MAX_NUM_OF_NEIGHBOURS_BLACK_HOLES];
 
   /*! Number of interactions in the force SELF and PAIR */
   int num_ngb_force;
 
   /*! List of interacting particles in the force SELF and PAIR */
-  long long ids_ngbs_force[MAX_NUM_OF_NEIGHBOURS_BLACK_HOLES];
+  int64_t ids_ngbs_force[MAX_NUM_OF_NEIGHBOURS_BLACK_HOLES];
 #endif
 
 } SWIFT_STRUCT_ALIGN;
