+++ src/hydro/AnarchyPU/hydro_part.h
@@ -80,7 +80,7 @@ struct xpart {
 struct part {
 
   /*! Particle unique ID. */
-  long long id;
+  int64_t id;
 
   /*! Pointer to corresponding gravity part. */
   struct gpart* gpart;
