+++ src/hydro/Gadget2/hydro_part.h
@@ -77,7 +77,7 @@ struct xpart {
 struct part {
 
   /* Particle ID. */
-  long long id;
+  int64_t id;
 
   /* Pointer to corresponding gravity part. */
   struct gpart* gpart;
@@ -177,10 +177,10 @@ struct part {
 
 #ifdef DEBUG_INTERACTIONS_SPH
   /*! List of interacting particles in the density SELF and PAIR */
-  long long ids_ngbs_density[MAX_NUM_OF_NEIGHBOURS];
+  int64_t ids_ngbs_density[MAX_NUM_OF_NEIGHBOURS];
 
   /*! List of interacting particles in the force SELF and PAIR */
-  long long ids_ngbs_force[MAX_NUM_OF_NEIGHBOURS];
+  int64_t ids_ngbs_force[MAX_NUM_OF_NEIGHBOURS];
 
   /*! Number of interactions in the density SELF and PAIR */
   int num_ngb_density;
