+++ src/tools.h
@@ -30,10 +30,10 @@
 
 void factor(int value, int *f1, int *f2);
 void density_dump(int N);
-void pairs_single_grav(double *dim, long long int pid,
+void pairs_single_grav(double *dim, int64_t pid,
                        struct gpart *restrict gparts, const struct part *parts,
                        int N, int periodic);
-void pairs_single_density(double *dim, long long int pid,
+void pairs_single_density(double *dim, int64_t pid,
                           struct part *restrict parts, int N, int periodic);
 
 void pairs_all_density(struct runner *r, struct cell *ci, struct cell *cj);
